requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.1
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
pymongo>=4.6
mongoengine>=0.27
bcrypt>=4.0

boto3>=1.34
Pillow>=10.0
tensorflow>=2.15
scikit-learn>=1.5
pandas>=2.2
numpy>=1.26
scipy>=1.11
torch>=2.4
torch-geometric>=2.5
sentence-transformers>=2.7
faiss-cpu>=1.7
tf-keras>=2.20
gunicorn>=21.2
requests>=2.31
matplotlib>=3.8
seaborn>=0.13
plotly>=5.18
streamlit>=1.38
tqdm>=4.66
joblib>=1.3
